pyproject.toml
[project]
name = "elt-dbt"
version = "0.1.0"
description = ""
authors = [
    {name = "rafael",email = "rafael.irvine@sempreceub.com"}
]
readme = "README.md"
requires-python = "^3.12.5"
dependencies = [
    "dbt-core (>=1.11.7,<2.0.0)",
    "dbt-snowflake (>=1.11.3,<2.0.0)",
    "dbt-postgres (>=1.10.0,<2.0.0)",
    "dbt-redshift (>=1.10.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "snowflake-sqlalchemy (>=1.9.0,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
